+++ src/KeyVault/KeyVault.Test/ScenarioTests/KeyVaultManagementTests.ps1
@@ -826,9 +826,10 @@ function Test-UpdateKeyVault {
         $vault = $vault | Update-AzKeyVault -DisableRbacAuthorization $false
         Assert-True { $vault.EnableRbacAuthorization } "5. EnableRbacAuthorization should be true"
 
+        # TODO: uncomment this afterwards. (Azure Client Tools Tenant does not have permission to turn off RBACAuth)
         #Set EnableRbacAuthorization false
-        $vault = $vault | Update-AzKeyVault -DisableRbacAuthorization $true
-        Assert-False { $vault.EnableRbacAuthorization } "6. EnableRbacAuthorization should be false"
+        #$vault = $vault | Update-AzKeyVault -DisableRbacAuthorization $true
+        #Assert-False { $vault.EnableRbacAuthorization } "6. EnableRbacAuthorization should be false"
 
         # Update Tags
         $vault = $vault | Update-AzKeyVault -Tag @{key = "value"}